+++ models/__init__.py
@@ -2,6 +2,7 @@ from .firewall import Firewall
 from .log_db import LogModel 
 from .ticket_db import TicketModel
 from .access_db import AccessModel 
-from .error_db import ErrorModel 
+from .error_db import ErrorModel
+from .ps_usage_db import UsageModel